process: split openInRoot and checkInodeDeviceMapping by build tag
Move openInRoot and the inode/device verification helper into
platform-specific files so coredump tests build and run on Darwin/ARM64:

Regression introduced by 234b685c.
